Conflicts:
	drake/systems/lcm/lcm_vector_interface_translator.h
	drake/systems/lcm/translator_between_lcmt_drake_signal.cc
	drake/systems/lcm/translator_lcmt_drake_signal.h

+++ drake/systems/plants/RigidBodySystem.cpp
@@ -56,7 +56,7 @@ size_t RigidBodySystem::getNumInputs(void) const {
 }
 
 size_t RigidBodySystem::getNumOutputs() const {
-  int n = getNumStates();
+  size_t n = getNumStates();
   for (const auto& s : sensors) {
     n += s->getNumOutputs();
   }
@@ -212,7 +212,7 @@ RigidBodySystem::StateVector<double> RigidBodySystem::dynamics(
   }
 
   if (tree->getNumPositionConstraints()) {
-    int nc = tree->getNumPositionConstraints();
+    size_t nc = tree->getNumPositionConstraints();
     const double alpha = 5.0;  // 1/time constant of position constraint
                                // satisfaction (see my latex rigid body notes)
 
@@ -260,7 +260,7 @@ RigidBodySystem::OutputVector<double> RigidBodySystem::output(
   DRAKE_ASSERT(getNumInputs() == u.size());
 
   y.segment(0, getNumStates()) << x;
-  int index = getNumStates();
+  size_t index = getNumStates();
   for (const auto& s : sensors) {
     y.segment(index, s->getNumOutputs()) = s->output(t, kinsol, u);
     index += s->getNumOutputs();
